deleted:    .idea/.gitignore 	deleted:    .idea/DaA_DETR.iml 	deleted:    .idea/git_toolbox_prj.xml 	deleted:    .idea/inspectionProfiles/Project_Default.xml 	deleted:    .idea/inspectionProfiles/profiles_settings.xml 	deleted:    .idea/misc.xml 	deleted:    .idea/modules.xml 	deleted:    .idea/vcs.xml 	modified:   datasets/Dataset_cls.py 	modified:   datasets/__pycache__/Dataset_cls.cpython-38.pyc 	modified:   datasets/__pycache__/coco_utils.cpython-38.pyc 	modified:   datasets/coco_utils.py 	modified:   main.py 	modified:   models/__pycache__/__init__.cpython-38.pyc 	modified:   models/__pycache__/backbone.cpython-38.pyc 	modified:   models/__pycache__/detr.cpython-38.pyc 	modified:   models/__pycache__/matcher.cpython-38.pyc 	modified:   models/__pycache__/position_embedding.cpython-38.pyc 	modified:   models/backbone.py 	modified:   util/__pycache__/__init__.cpython-38.pyc 	modified:   util/__pycache__/box_ops.cpython-38.pyc 	modified:   util/__pycache__/distributed_utils.cpython-38.pyc 	modified:   util/__pycache__/misc.cpython-38.pyc 	modified:   util/__pycache__/position_ops.cpython-38.pyc 	modified:   util/__pycache__/train_utils.cpython-38.pyc 	modified:   util/__pycache__/util.cpython-38.pyc 	modified:   util/misc.py 	modified:   util/train_utils.py

+++ util/misc.py
@@ -289,7 +289,7 @@ class NestedTensor(object):
 
     def to(self, device):
         # type: (Device) -> NestedTensor # noqa
-        cast_tensor = self.tensors.to(device)
+        cast_tensor = self.tensors.to(device).float()
         mask = self.mask
         if mask is not None:
             assert mask is not None
